feat(.bluemix): improve UX and docs of the guided install (#329)
Combination of 6 squashed commits:
chore(.bluemix/nls): expand and improve toolchain descriptions of tools and items

refactor(.bluemix/deploy): split table to 3 coloums to give fields space and move label

feat(.bluemix): add missing info urls and placeholders to repo items

feat(.bluemix): move link existing doi toolchain to optional section to match CI pipeline

The link to existing DOI toolchain has a problem in the guided installer,
if the team doesn't have a DOI on another toolchain to provide to the
step, then this becomes a blocking issues, since the guided installer
won't allow you to continue forward until you provide either a valid
toolchain id, or a fake one. When looking at the CI pipeline, we can see
that this same setup step is listed in optional, since you can also
setup a DOI on the toolchain itself and use that for the link later. CD
pipeline should continue that behavior, or enforce in its setup.md that
an existing Toolchain with DOI is an absolute requirement for CD.

docs(setup): reformat and grammer/spell check

docs(setup): move sections of the tool ints. to match the guided install's optional steps

Co-authored-by: PADRAIC EDWARDS <padraic.edwards@ie.ibm.com>